refactor: migrate transactions and users to new toast
